Move JstSessionBuilder object to appropriate module
