+++ app/concepts/api/v1/auth/operation/sign_up.rb
@@ -19,7 +19,7 @@ module API::V1::Auth::Operation
     end
 
     def generate_auth_token(ctx, model:, **)
-      session = JwtSessionBuilder.new.call(user_id: model.id)
+      session = JwtSession::Build.new.call(user_id: model.id)
       ctx['result'] = session.login
       ctx['operation_status'] = :success
     end
